Move isTypeIncluded to PrjTypes and deprecate old helpers
Refactored code by relocating isTypeIncluded from Helper to PrjTypes for better cohesion. Deprecated several Helper methods, marking them for future removal.
